% Generated by roxygen2 (4.0.2): do not edit by hand
\name{as.AggregatedConcentrations.SpatialPointsDataFrame}
\alias{as.AggregatedConcentrations.SpatialPointsDataFrame}
\alias{ggplot.AggregatedConcentrations}
\alias{spplot.AggregatedConcentrations}
\title{Visualization}
\usage{
as.AggregatedConcentrations.SpatialPointsDataFrame(from)

spplot.AggregatedConcentrations(obj, ...)

\method{ggplot}{AggregatedConcentrations}(data, bounds, ...)
}
\arguments{
\item{from}{an AggregatedConcentrations object}

\item{...}{further arguments to spplot}

\item{x}{an AggregatedConcentrations object, obtained by calling aggregate() on the results of a modeling run}
}
\value{
a SpatialPointsDataFrame
}
\description{
Plot aggregate concentrations using spplot or ggplot2.

Visualization
}
\keyword{ggplot2}
\keyword{spplot}

