\name{as}
\alias{as}
\title{Re-bind results obtained from \code{aggregate(x)} to a SpatialPointsDataFrame.}
\arguments{
  \item{from}{an AggregatedConcentrations object}
}
\value{
  a SpatialPointsDataFrame
}
\description{
  Re-bind results obtained from \code{aggregate(x)} to a
  SpatialPointsDataFrame.
}
\keyword{model}
\keyword{predict}

