\name{as.SpatialLines.FreeFlowLinks}
\alias{as.SpatialLines.FreeFlowLinks}
\title{Extract the geometry from a FreeFlowLinks object.}
\usage{
  as.SpatialLines.FreeFlowLinks(links)
}
\arguments{
  \item{links}{a FreeFlowLinks object}
}
\value{
  a SpatialLines object
}
\description{
  Returns a SpatialLines object.
}
\keyword{links}

