% Generated by roxygen2 (4.0.2): do not edit by hand
\name{links}
\alias{links}
\alias{links.default}
\alias{meteorology}
\alias{meteorology.default}
\alias{pollutant}
\alias{pollutant.default}
\alias{receptors}
\alias{receptors.default}
\alias{terrain}
\alias{terrain.default}
\title{Caline3Model accessors}
\usage{
links(model, ...)

\method{links}{default}(model, ...)

receptors(model, ...)

\method{receptors}{default}(model, ...)

meteorology(model, ...)

\method{meteorology}{default}(model, ...)

pollutant(model, ...)

\method{pollutant}{default}(model, ...)

terrain(model, ...)

\method{terrain}{default}(model, ...)
}
\arguments{
\item{model}{a Caline3Model object}

\item{...}{further arguments}
}
\description{
Inspect model elements from a Caline3Model.

Caline3Model accessors

Caline3Model accessors

Caline3Model accessors

Caline3Model accessors

Caline3Model accessors

Caline3Model accessors

Caline3Model accessors

Caline3Model accessors

Caline3Model Methods
}

