% Generated by roxygen2 (4.0.2): do not edit by hand
\name{extract_segments}
\alias{extract_segments}
\alias{extract_segments.Lines}
\alias{extract_segments.SpatialLines}
\alias{extract_segments.default}
\title{Extract segments from a polyline or polylines}
\usage{
extract_segments(x, ...)

extract_segments.default(x, suffixes = c(".start", ".end"))

extract_segments.Lines(x, ...)

extract_segments.SpatialLines(x, ...)
}
\arguments{
\item{x}{Lines or SpatialLines object}
}
\value{
matrix of coordinates
}
\description{
Extract segments from a polyline or polylines
}

