\name{FortranArrays}
\alias{FortranArrays}
\title{Construct a FortranArrays object.}
\usage{
  FortranArrays(as.real = as.single, ...)
}
\arguments{
  \item{as.real}{function to convert numeric arguments to
  appropriate type (single or double)}

  \item{...}{named arguments}
}
\value{
  a list object
}
\description{
  Construct a FortranArrays object.
}

