% Generated by roxygen2 (4.0.2): do not edit by hand
\name{centerlines}
\alias{centerlines}
\alias{centerlines.FreeFlowLinks}
\title{FreeFlowLinks accessors}
\usage{
centerlines(obj, ...)

\method{centerlines}{FreeFlowLinks}(obj, ...)
}
\arguments{
\item{obj}{a FreeFlowLinks object}

\item{...}{further arguments}
}
\description{
Inspect attributes of a FreeFlowLinks object.

FreeFlowLinks accessors
}

