% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ggplot.Caline3Model}
\alias{ggplot.Caline3Model}
\title{Visualization}
\usage{
\method{ggplot}{Caline3Model}(data, ...)
}
\arguments{
\item{data}{a Caline3Model object}
}
\description{
Plot model elements using ggplot2.
}
\keyword{ggplot}
\keyword{ggplot2}

