% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ggplot.Meteorology}
\alias{ggplot.Meteorology}
\title{ggplot.Meteorology}
\usage{
\method{ggplot}{Meteorology}(data, ...)
}
\arguments{
\item{...}{other arguments}

\item{x}{a Meteorology object}
}
\description{
Plots a wind rose using ggplot2.
}
\keyword{meteorology}

