% Generated by roxygen2 (4.0.2): do not edit by hand
\name{lines.FreeFlowLinks}
\alias{lines.FreeFlowLinks}
\title{lines}
\usage{
\method{lines}{FreeFlowLinks}(x, ...)
}
\arguments{
\item{x}{a FreeFlowLinks object}

\item{...}{other arguments}
}
\description{
Plot a FreeFlowLinks object.
}
\keyword{links}

