% Generated by roxygen2 (4.0.2): do not edit by hand
\name{Meteorology}
\alias{Meteorology}
\alias{Meteorology.ISCFile}
\alias{Meteorology.character}
\alias{Meteorology.data.frame}
\title{Meteorology}
\usage{
Meteorology(x, ...)

\method{Meteorology}{data.frame}(x)

\method{Meteorology}{character}(x)

\method{Meteorology}{ISCFile}(x, use = c("urban", "rural"))
}
\arguments{
\item{x}{a filename, \link{ISCFile} object, or data.frame}

\item{use}{whether to use urban or rural mixing heights}
}
\value{
a Meteorology object (essentially a \link{data.frame})
}
\description{
Construct a Meteorology object, generally from a file containing ISC-formatted records.
}
\keyword{meteorology}

