% Generated by roxygen2 (4.0.2): do not edit by hand
\name{Pollutant}
\alias{Pollutant}
\title{Construct a Pollutant object.}
\usage{
Pollutant(name, molecularWeight, settlingVelocity = 0,
  depositionVelocity = 0)
}
\arguments{
\item{name}{TODO}

\item{molecularWeight}{TODO}

\item{settlingVelocity}{TODO}

\item{depositionVelocity}{TODO}
}
\value{
a Pollutant object
}
\description{
Use to override the default pollutant in \code{\link{Caline3Model}}.
}
\keyword{model}
\keyword{pollutant}

