% Generated by roxygen2 (4.0.2): do not edit by hand
\name{predict.Caline3Model}
\alias{predict.Caline3Model}
\title{predict.Caline3Model}
\usage{
\method{predict}{Caline3Model}(object, units = "ppm")
}
\arguments{
\item{object}{a \code{\link{Caline3Model}} object}

\item{units}{one of 'mg/m^3', 'ug/m^3', or 'ppm'}
}
\value{
matrix of predicted values
}
\description{
predict.Caline3Model
}
\keyword{model}
\keyword{predict}

