\docType{package}
\name{Rcaline}
\alias{Rcaline}
\alias{Rcaline-package}
\title{Rcaline}
\description{
  Rcaline provides an interface to Fortran implementations
  of the CALINE model family. These steady-state, Gaussian
  dispersion models are used to predict aerosol
  concentrations downwind from mobile emission sources such
  as highway traffic.
}
\author{
  David Holstius \email{david.holstius@berkeley.edu}
}
\references{
  Benson, P. (1979) CALINE-3: A versatile dispersion model
  for predicting air pollution levels near highways and
  urban streets. Federal Highway Authority report
  FHWA/CA/TL-79/23, California DOT, Sacramento (1979).
}
\keyword{CALINE}
\keyword{CALINE3}
\keyword{disperson}
\keyword{environmental}
\keyword{modeling}
\keyword{package}
\keyword{traffic}

