% Generated by roxygen2 (4.0.2): do not edit by hand
\name{Receptors}
\alias{Receptors}
\alias{Receptors.SpatialPoints}
\alias{Receptors.SpatialPointsDataFrame}
\alias{Receptors.matrix}
\title{Receptor locations}
\usage{
Receptors(obj, ...)

\method{Receptors}{SpatialPointsDataFrame}(obj, z = 1.8,
  check.projection = TRUE)

\method{Receptors}{SpatialPoints}(obj, z = 1.8, check.projection = TRUE)

\method{Receptors}{matrix}(obj, z = 1.8)
}
\arguments{
\item{obj}{a SpatialPoints or SpatialPointsDataFrame object, or a two-dimensional matrix of coordinates}

\item{z}{elevation(s) above ground, in meters}

\item{check.projection}{checks to see that obj has an associated projection}
}
\value{
SpatialPoints* object
}
\description{
Construct a set of receptors. All coordinates must be in meters!
}
\seealso{
ReceptorGrid ReceptorRings
}
\keyword{receptors}

