% Generated by roxygen2 (4.0.2): do not edit by hand
\name{Rectangle}
\alias{Rectangle}
\alias{Rectangle.Spatial}
\alias{Rectangle.matrix}
\title{Rectangle}
\usage{
Rectangle(obj, ...)

Rectangle.matrix(obj)

Rectangle.Spatial(obj)
}
\arguments{
\item{obj}{a matrix or Spatial* object}
}
\value{
a SpatialPolygons object with one Polygon
}
\description{
Create a Rectangle object.

Rectangle.matrix

Rectangle.Spatial
}
\keyword{Rectangle}

