\name{sample.rows}
\alias{sample.rows}
\title{Take a random sample of rows from a data.frame.}
\usage{
  sample.rows(x, p)
}
\arguments{
  \item{x}{a data.frame}

  \item{p}{the fraction of rows to sample (between 0 and
  1)}
}
\value{
  a subset of the original data.frame
}
\description{
  Take a random sample of rows from a data.frame.
}
\keyword{data.frame}
\keyword{sample}

