\name{suppressOutput}
\alias{suppressOutput}
\title{Suppress all output while evaluating an expression.}
\usage{
  suppressOutput(expr)
}
\arguments{
  \item{expr}{an \link{expression}}
}
\value{
  the value of expr
}
\description{
  Suppress all output while evaluating an expression.
}
\seealso{
  suppressMessages suppressWarnings
}

