% Generated by roxygen2 (4.0.2): do not edit by hand
\name{Terrain}
\alias{Terrain}
\title{Construct a Terrain object.}
\usage{
Terrain(surfaceRoughness)
}
\arguments{
\item{surfaceRoughness}{}
}
\value{
a Terrain object
}
\description{
Construct a Terrain object.
}
\keyword{model}
\keyword{terrain}

