\name{units}
\alias{units}
\alias{units<-}
\title{Get the "units" attribute on a numeric vector.}
\usage{
  units(x)

  units(x) <- value
}
\arguments{
  \item{x}{numeric vector}

  \item{value}{replacement value}
}
\description{
  Get the "units" attribute on a numeric vector.

  Set the "units" attribute on a numeric vector.
}
\keyword{units}

