\name{lines.FreeFlowLinks}
\alias{lines.FreeFlowLinks}
\title{Plot a FreeFlowLinks object using base graphics.}
\usage{
  lines.FreeFlowLinks(x, ...)
}
\arguments{
  \item{x}{a FreeFlowLinks object}

  \item{...}{other arguments}
}
\description{
  Works just like \code{\link{lines}}.
}
\keyword{links}

