\name{Parameters}
\alias{Parameters}
\title{Construct a Parameters object.}
\usage{
  Parameters(surfaceRoughness, averagingTime = 60,
    settlingVelocity = 0, depositionVelocity = 0)
}
\arguments{
  \item{surfaceRoughness}{in centimeters}

  \item{averagingTime}{in minutes}

  \item{settlingVelocity}{of the pollutant}

  \item{depositionVelocity}{of the pollutant}
}
\value{
  a Parameters object
}
\description{
  Construct a Parameters object.
}

