\name{predict.Caline3Model}
\alias{predict.Caline3Model}
\title{predict.Caline3Model}
\usage{
  predict.Caline3Model(object, units = "ppm")
}
\arguments{
  \item{object}{a \code{\link{Caline3Model}} object}

  \item{units}{one of 'ppm', 'ug/m3', or 'mg/m3'}
}
\value{
  matrix of predicted values
}
\description{
  predict.Caline3Model
}
\keyword{model}
\keyword{predict}

