\name{Rcaline}
\alias{Rcaline}
\alias{Rcaline-package}
\docType{package}
\title{Rcaline}
\description{
Rcaline provides an interface to Fortran implementations of the CALINE model family. These steady-state, Gaussian dispersion models are used to predict aerosol concentrations downwind from mobile emission sources such as highway traffic.
}
\details{
\tabular{ll}{
Package: \tab Rcaline\cr
Type: \tab Package\cr
Version: \tab 0.8\cr
Date: \tab 2011-01-11\cr
License: \tab FILE\cr
LazyLoad: \tab yes\cr
}

}
\author{
David Holstius <david.holstius@berkeley.edu>
}
\references{
Benson, P. (1979) CALINE-3: A versatile dispersion model for predicting air pollution levels near highways and urban streets. Federal Highway Authority report FHWA/CA/TL-79/23, California DOT, Sacramento (1979).
}
\keyword{ 
package
atmospheric modeling 
dispersion
CALINE
CALINE3
}
