\name{Receptors}
\alias{Receptors}
\title{Promote a SpatialPoints* object to a Receptors object.}
\usage{
  Receptors(spobj, elevation)
}
\arguments{
  \item{spobj}{an object of class SpatialPoints or
  SpatialPointsDataFrame}

  \item{elevation}{elevation above ground, in meters}
}
\value{
  object of class Receptors
}
\description{
  Promote a SpatialPoints* object to a Receptors object.
}
\keyword{receptors}

