\name{ReceptorSurface}
\alias{ReceptorSurface}
\title{Promote a SpatialPoints* object to a ReceptorSurface object.}
\usage{
  ReceptorSurface(spobj, elevation)
}
\arguments{
  \item{spobj}{an object of class SpatialPoints or
  SpatialPointsDataFrame}

  \item{elevation}{elevation above ground, in meters}
}
\value{
  object of class ReceptorSurface
}
\description{
  Promote a SpatialPoints* object to a ReceptorSurface object.
}
\keyword{receptors}

