\name{bbox.Rectangle}
\alias{as.SpatialPolygons.Rectangle}
\alias{as.array.Rectangle}
\alias{as.data.frame.Rectangle}
\alias{bbox.Rectangle}
\title{bbox.Rectangle}
\usage{
  bbox.Rectangle(obj)

  as.SpatialPolygons.Rectangle(from)

  as.array.Rectangle(from)

  as.data.frame.Rectangle(from)
}
\description{
  bbox.Rectangle

  as.Rectangle.SpatialPolygons

  as.array.Rectangle

  as.data.frame.Rectangle
}

